Copy faster-polymerize to a fresh fork of master.
* utils/_fastsums.pyx implements the sum_monomers() inner loop in Cython
for polymerize using typed Memoryviews. It turns out to be a bit tricky to
ensure Cython code actually runs faster than Python code -- sensitive to
details like testing that an index is out of range using
`if i < 0 or i >= max` instead of `if i not in xrange(max)`.
* Split out the profiling code from polymerize.py to
tests/utils/profile_polymerize.py where it can tweak the sys.path as a way
to workaround a Python import problem.
* Add unit tests for the added Cython code.
* In polymerize.py, define a no-op version of @profile when needed, so we
don't have to modify the source code to enable line profiling.
* Change Makefile to find the pyenv virtualenv-prefix rather than
hard-coding it.
* Use numpy.testing.assert_equal() to make test_polymerize.py simpler and
more informative.