polymerize() docstring and array definition comments
* This function juggles a bunch of arrays, and Python code doesn’t
distinguish scalars from arrays or explain the array indexes and
contents. So add comments to do that.

* ndarray.astype() makes a copy so there’s no need to make another.
Well, passing in copy=False allows it to return the input array if
possible. Pass in copy=True here to make it clear even though that’s
the default.

Add to a test case to verify that this function doesn’t modify the
caller’s array.

* PyCharm points out that the import from `_build_sequences` is
unused here, but other code relies on this file reexporting those
functions, so use __all__ to make it official.