polymerize() docstring and array definition comments (#116)
* polymerize() docstring, using NumPy parameter style.
This function juggles a bunch of arrays, and Python code doesn’t
distinguish scalars from arrays or explain the array indexes and
contents. So add comments to do that.

* `ndarray.astype()` makes a copy so there’s no need to make another.
Well, passing in `copy=False` allows it to return the input array if
possible. Pass in `copy=True` here to make it explicit (it’s
the default).

* Add to a test case to verify that polymerize() doesn’t modify the
caller’s array.

* This module doesn't use `_build_sequences` itself, but it reexports
the functions defined in that private (_xxx name) module, insulating
callers from knowing they're in Cython. Use `__all__` to make that
official to tools.

* Switch to explicit relative imports to avoid potential problems.
(Absolute imports also work but there are warnings about importing a
private `_name` that way.)
* Rename `@profile`’s parameter from `function` to `func` since
there’s a built-in type `function`.

* Decorate polymerize() with @profile in the
performance test rather than the production code.

* Put all imports at the top.