Re: #392 avoid seqfault and parallelization problems on macOS (#394)
Avoid `fork`ing processes via `multiprocessing` on macOS to avoid the segfault and serialization bugs, except the Agent Shepherd which is handled in PR #389 and the parallel Fitter which seems to work on macOS.

See Issue #392.

`parallelization.cpus()`:
* Limit to 1 core on macOS unless the caller overrides that safety check.
* Override that safety check in the parallel Fitter since it seems to work. It still only goes parallel when given a `--cpus` command line arg.
* Simplify callers by handling all the limits and the warning message. This avoids multiple code paths having to apply the same limits.

Also:
* Fix some PyCharm warnings, mainly commenting out unused assignments (presumably ones w/o side-effects).
* Adopt print_function while editing some files.
* Tweak a few print statements.

---

Yes, ultimately we should find an alternative for multiprocessing on macOS. Maybe Python 3 multiprocessing "spawn" mode. Failing that, `exec` sounds reasonable for the analysis classes as for the Agent Shepherd, and maybe the Fitter could use `exec` + sending pickled state in via a pipe.