Enable sending agent daughter cell info between servers
To distribute cell agents across multiple machines, the daughter cell inherited state data must get transmitted from one server to another (unless there's a shared file server). It's a 250KB BLOB (binary large object) and unsuitable to string-escape in JSON in the Kafka message payload, a print statement, or a command-line arg to the new agent.

So this changes our Kafka message payload format to an extensible list of chunks. The BLOBs get sent as separate chunks, and the Shepherd puts them into temp files and passes the filenames on the command line to the new cell agent.
* NOTE: This changes `inherited_state_path` from the directory path containing an implied filename to the actual file path, and changes the path from `simOut/Daughter1/inherited_state.cPickle` to `simOut/Daughter1_inherited_state.cPickle` et al. I updated `diff_simouts.py` but it's not smart about the diffs between old and new simOut dirs.
* NOTE: This requires the paired `multipart-msg` PR in the `shepherd` repo. The `shepherd` and `lens` repos could be merged to share the agent message encoder/decoder et al.
* Improved agent command line interface usability via config defaults, arg checks, updated --help, and updated README. `argparse.RawDescriptionHelpFormatter` enables manual line breaks in the help text.
* Outer agent now says why it's exiting after exceeding `max_time`, and Lattice has a longer `max_time` to reach cell divisions.
* The EnvironmentSimulation interface has default values for clarity and fewer PyCharm warnings.
* Tuned the agent logging a little, e.g. print the blob count rather than content.
    --> cell-receive ENVIRONMENT_UPDATE [ecoli ca63343d-029a-11e9-a23f-00e04c68076e]: {...}
   That's the Kafka topic, agent event, [agent type and ID], {message dict}.