install openblas v0.3.5 instead of 0.3.3 (#508)
* Install openblas v0.3.5 instead of 0.3.3 since that fixes threading bugs which cause inconsistent results and brings performance improvements.
* Apply a one-layer technique to keep openblas’ and GLPK’ intermediate files out of the Docker images.
* Set multiple ENV variables in one step to save layers, which should save build time.
* This brought the wholecell Docker image down another 390MB to 1.13GB.
* Move `requirements.txt` back to `wcEcoli/`.
* Give each `Dockerfile` its own directory.

TODO: Does the `gcloud builds submit` command need updating to identify the root directory and the Dockerfile? The documentation doesn't say how to do that.

Q. Does it need a `.gcloudignore` file? Do all the build commands find the one, project level `.dockerignore` file?

With openblas v0.3.5 now, runSim on macOS took about 24% (1 dB) longer in-container-in-Docker-VM than out-of-container. (I'll compare them on linux on GCE where there won't be a VM layer.) Also it ended closer to the simulation time and dry mass weight as it did out-of-container, but still not the same time.