simplify
The Task constructor doesn't need the `or []` now that it accepts Iterables. All that'd be simpler if Python default arguments weren't nasty with mutable values like `[]`.

Don't print a comma between the two `--dump` filenames. This makes it easier to paste the names into a command line that opens them in an editor.