test python exception handling in a workflow (#636)
* switch the WF builder from 'STDOUT' to '>'
* test error handling in a Sisyphus command
* Add a demo task that applies `wc` to a missing input file. This tests that Sisyphus logs the return code and uploads the stderr file to storage even though the command hit an error, which is of course exactly when you want the stderr output.
* fix spaces in multiline strings
* make lots of log filename sort nicely by using 3-digit variant, seed, gen, and cell numbers.
* add a Gaia test to the "demo" workflow for python exception handling

In issue #635 , when the simulation task raised an exception and exited, Sisyphus uploaded the stderr (including stack trace) then Gaia kept rerunning the failed task on other workers. This adds a small test case to see if it's just the exception that causes that behavior and to speed up debugging.