lens-interface==0.0.27
Lens now has a small pip `lens-interface==0.0.27` to install into sim projects like wcEcoli, enough to let them play as actors.

This is now separate from the pip `wholecell-lens` which has the Lens application et al to run in a separate process with a separate pyenv -- and optionally a different version of Python, which aids porting.