Generalize Behavior Metrics
Based on review comments from @1fish2, we might want to use the output
of these metrics for more than just unit testing. For example, we might
create plots of these metrics over time or run checks only over the
running averages. To accommodate this use case, this commit removes the
unit testing framework and instead stores results in a Pandas DataFrame
for other classes to consume.

This commit also changes the BehaviorMetrics class to take in the
simulation output directory and configuration file paths as constructor
arguments for more general usability.

Example Output:

```
$ python prototypes/behavior_metrics/behavior_metrics.py
              metric  expected_min  expected_max     value   pass
0        proteinMass           1.9           2.1  1.921763   True
1           rRnaMass           1.9           2.1  1.971384   True
2            dnaMass           1.9           2.2  2.110126   True
3  smallMoleculeMass           1.9           2.1  1.949451   True
4           tRnaMass           1.9           2.1  1.988803   True
5           mRnaMass           1.9           2.1  1.837701  False
```

Addresses some of the review comments from @1fish2 and
@prismofeverything. Fixes part of #675.