Allow Multiple Modes Per Metric
Adjust config file JSON schema to allow for a given metric to be
analyzed with multiple modes. This allows us to more easily specify
bounds for a metric's mean, stdev, end-start ratio, min, max, etc.
without defining a whole new metric for each.

Here is an example of the resulting dataframe. Note that now not all the
tests pass to illustrate failure cases.
```
              metric             mode  expected_min  expected_max     value   pass
0        proteinMass  end_start_ratio           1.9           2.1  1.921763   True
1           rRnaMass  end_start_ratio           1.9           2.1  1.971384   True
2            dnaMass  end_start_ratio           1.9           2.1  2.110126  False
3  smallMoleculeMass  end_start_ratio           1.9           2.1  1.949451   True
4           tRnaMass  end_start_ratio           1.9           2.1  1.988803   True
5           mRnaMass  end_start_ratio           1.9           2.1  1.837701  False
```

Fixes part of #675.