Create Function to Order Operations
To allow operations (modes) to depend on each other, we need to order
them such that they for any pair of operations A and B where B depends
on A, A is evaluated before B. We do this using the topological sort
function in networkx in the `order_operations` function.

Fixes part of #724.