Add Unit Test for Detecting Operation Graph Cycles
If the dependencies specified in the arguments of operations form a
cycle (e.g. if A takes B as an argument, B takes C, and C takes A), we
cannot perform the operations. Such a cycle is detected when we try
performing a topological sort on the dependency graph generated by the
operation arguments. This commit adds a unit test to check that an
informative exception is raised if such a cycle is found.

Fixes part of #724.