Fix test_initial_conditions.Test_InitialConditions
The unittest `test_initial_conditions.Test_InitialConditions` was not
updated when the `n_max_replisomes` parameter was introduced to
`determine_chromosome_state` in
8668f67c5c0b5a35245770dead6e021916b922db, which caused test failures in
pytest. This commit fixes the unit test by passing in a large constant
to `n_max_replisomes` so that the tests run as expected.

Fixes part of #45.