Update Documentation for Pytest from Nosetests
For the most part, this is just replacing `nosetests` with `pytest` in
documentation files and comments. One exception is
`wholecell/tests/utils/test_library_performance.py` where a line that
relies on nosetests attributes is removed because running all the tests
in the file should have the same effect.

Fixes part of #45.