Replace Deprecated BaseException.message Calls
In unit tests, we were frequently checking error messages like this:

```
with self.Raises(Exception) as context:
    ...
    self.assertEqual(context.exception.message, "...")
```

However, the `context.exception.messasge` attribute has been deprecated
in Python 2.6 and will be removed in Python 3 (See
[PEP 352](https://www.python.org/dev/peps/pep-0352/#id8)). This commit
replaces this code with the following structure:

```
with self.RaisesRegexp(
    Exception,
    "^...$",
):
    ...
```

This resolves the last of the pytest warnings.

Fixes part of #45.