Add Support for Units in Loaded Data
When loading data, the `units` key can now be used to specify units to
apply to the loaded data. The value of `units` can be either a string
like `m*g/L` that describes the units or a name to import, e.g.
`wholecell.utils.units.L`. This allows for easier handling of data where
units are important to track.

Fixes part of #733.