Add Limiting Metabolites Metric
Add a metric for metabolites that become limiting over the course of the
simulation. To support this metric, a new kind of test is added for
checking metrics of type Set. In the output data frame, the min and max
are set to None, and the metric value is the produced Set. This will
require that for plotting, non-numeric metrics be excluded.

Resolves part of #733.