compute task dependencies
Compute task-to-task dependencies in the workflow builder so its caller doesn't have to do that to create a FireWorks workflow. This also provides up-front checking for conflicting outputs and unfulfilled inputs.

It computes task dependencies from their inputs & outputs, which a FireWorks workflow still needs to copy the files to/from GCS.