Decompose proteinCountsValidation Analysis Script
The analysis script includes logic that will be repurposed for the
proteome correlation behavior metric. To allow this re-use, this commit
decomposes the analysis code to isolate the common logic into utility
functions in the analysis script. Later, these functions should be moved
into a utility file.

Fixes part of #733.