A Python 2 EOL issue: virtualenv 20.0.4 requires a newer pathlib2
pip pathlib2 2.3.2 changed to a new syntax for conditional dependencies in `setup.py`, which broke something without even bumping the major or minor version. That's a Bozo no no.

`pathlib2` is a backport of Python 3 `pathlib` to Python 2.

Release 2.3.3 fixed that, and release 2.3.5 seems to be the final for this library since "Python 2 is no longer supported." The GitHub repo is now archived.

https://github.com/mcmtroffaes/pathlib2/issues/46

The current `virtualenv` 20.0.4 requires `pathlib2<3,>=2.3.3`, so update to the final `pathlib2` release 2.3.5. The alternative is to freeze on an older version of `virtualenv`. It's new territory either way.