update to OpenBLAS 0.3.9
In case the bug fixes matter, update OpenBLAS to 0.3.9 inside the Docker container and in the create-pyenv instructions.

I retested the problem with AVX2 instructions in Docker Desktop for Mac (https://github.com/xianyi/OpenBLAS/issues/2244) with the latest OpenBLAS and Docker Desktop and filed the Issue in the Docker repo this time, https://github.com/docker/for-mac/issues/4576

On macOS outside of Docker, `brew install openblas` now installs 0.3.9 . We no longer need to compile it from source.

Use Python 3.8.3 in the test workflows.