Migrate wcEcoli in vivarium to Use Hierarchy
The previous model that could simulate wcEcoli as a process in vivarium
no longer works now that we have changed vivarium to use hierarchy. This
commit migrates the old model to work with hierarchy.

The model can now be run as an experiment by executing
environment/lattice_experiment.py, but we do not yet support modeling
division.