Update Vivarium and Stochastic Arrow Requirements
This PR replaces the wholecell-lens requirement with wholecell-vivarium, upgrades stochastic-arrow requirement to its latest version, and freezes on the new dependencies.

Upgrading vivarium will also facilitate the migration to Python 3.

* Upgrade stochastic-arrow

Between 0.2.0 and 0.4.1, the interface for stochastic-arrow changed to
specify rates in evolve_state instead of the constructor. This commit
upgrades the stochastic-arrow dependency in the requirements file and
fixes stochastic-arrow usages to match the new interface.

* Transition from Lens to Vivarium

Now that Lens has been replaced with Vivarium, this commit changes the
code to use Vivarium instead of Lens. Vivarium includes most of the Lens
functionality, so this commit is trivial name change.

* Bump vivarium version

* Indent with tabs instead of spaces

* Fix arrow test for upgraded interface

* remove the pygame requirement and let Jenkins test using virtualenv wcEcoli2-staging

* comment out the emitter code for now

* restore ecoli-pull-request.sh to wcEcoli2

* specify the additional pips installed by wholecell-vivarium