use different km.cPickle filenames
Switch from km.cPickle to km2.cPickle and km3.cPickle, depending on Python version, so it won't throw an exception trying to read an incompatible file.

Switching from numpy `'a'` fields to `'U'` makes the contents incompatible with older code. We could make the Python 2 and 3 files compatible but it's not worth the work for this cache file.