simpler pyenv setup via default OpenBLAS
Current binary "wheels" to install numpy and scipy now embed a fine version of OpenBLAS, 0.3.9.dev (anything 3.7+ is fine), so we no longer need to install an OpenBLAS package or compile it from source, create `~/.numpy-site.cfg`, or install numpy and scipy pips with `--no-binary`. This is much simpler and faster, like 11 seconds to install those two rather than 15-30 minutes.

This commit updates the instructions in `py3_requirements.txt`.

* TODO: Update the .md files, wiki pages.
* TODO: Update the virtualenv on Sherlock and test that it has a good openblas.