workaround `lpad webgui` bug
`lpad webgui` fails, raising

    TypeError: cannot pickle '_thread.lock' object

in Python 3.8.0 - 3.8.5. It's a bug in either Python 3.8 or FireWorks. It works fine in Python 3.7.

As a workaround, run `lpad webgui -s`. That's "server mode" which doesn't open a web page (or fail trying) but in iTerm2 you can Cmd-click the link it prints `http://127.0.0.1:5000` or otherwise paste that URL into the browser.

This workaround requires the new FireWorks release to avoid the error:

    Gunicorn is required for server mode. Install using `pip install gunicorn`.

even though we have already installed gunicorn.