Set Emission Frequency to Emit ~100 timepoints
For long simulations, we don't want to emit every second, because this
generates way more data than we need. The surplus of data makes analysis
much more difficult and resource-intensive. Instead, we now set the
emission frequency so that we emit ~100 timepoints (or every second if
we run for <100 seconds).