Deliver Pulses Across Field, Not At Single Point
When delivering antibiotic pulses to a single point in the field, I
found that it very quickly diffused throughout the environment. This
occurred just as quickly as antibiotics flowed into the cells, making
the single-point-delivery less useful. Further, the snapshots plot
invariably showed no changes because the concentrations quickly dropped
to near zero as the antibiotic diffused (relative to the amount
initially deposited). Therefore, we now provide antibiotic uniformly
across the environment with each pulse.