Added computation, listener, and analysis for mass by compartment (#967)
* listener and analysis for compartment masses

* cleaning up compartment mass analysis code before PR

* Changes in PR comments made, including changing self._compartment_masses
to a matrix and removing the for loop from the calculation.
Additionally, changes were made to include the compartment labeled [s],
corresponding to "CCO-FLAGELLUM", which was previously missing. Unique
molecules still needs to be changed to calculate compartment_masses in the
same way as bulk_molecules.

* Compartment submasses added to unique molecules, where all unique
molecule mass is added to the cytosol. A few code style fixes
added as well.

* Indexing changed to be able to use CCO id's instead of compartment
abbrevs.

* Style changes for second PR review.