Fix #996
This should fix #996 by putting more cross-checks in the Km counts cache, but I don't have a test case for that.

Ideally we'd factor out this part of the very long `setKmCooperativeEndoRNonLinearRNAdecay()` function and add test cases for it.

Q. Should `arrays_differ()` set the absolute and/or relative tolerances?

Adding to the cache file grew it from 37K to 110K. That should be fine but